:bug: Move all python files to single folder. Update `setup.sh`` to set python path
