+++ setup.sh
@@ -1,11 +1,9 @@
-#!/bin/bash
-
 # Run flashcards pipeline
-python -m src.backend.RAG.flashcards_pipeline
-echo "flascards pipeline executed!"
+PYTHONPATH=./src python3 src/backend/flashcards_pipeline.py
+echo "flashcards pipeline executed!"
 
 # Run knowledge base pipeline
-python -m src.backend.RAG.knowledge_base_pipeline
+PYTHONPATH=./src python3 src/backend/knowledge_base_pipeline.py
 echo "knowledge pipeline executed!"
 
 # Run another script if needed
