+++ src/backend/Flashcards_llm_ollama.py
@@ -1,8 +1,5 @@
-from langchain_ollama import OllamaLLM
-from langchain_core.prompts import ChatPromptTemplate
-from .template import create_template, create_question_template, create_answer_template
-from .vectorstore import load_vectorstore
-from .question_generator_base import QuestionGeneratorBase
+from template import create_answer_template
+from question_generator_base import QuestionGeneratorBase
 
 
 class FlashCards(QuestionGeneratorBase):
