+++ src/backend/knowledge_base.py
@@ -1,7 +1,7 @@
 import chromadb
 import os
-from .local_loader import load_documents
-from .splitter import smart_split_documents
+from local_loader import load_documents
+from splitter import smart_split_documents
 import torch
