+++ src/backend/main.py
@@ -2,11 +2,11 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from endpoints import router
 import os
-from RAG.rag_pipeline import insert_documents
+from rag_pipeline import insert_documents, CHROMA_DIR
 from ait_logger import logger
 
 
-chroma_path = os.path.join(os.path.dirname(__file__), "RAG", "chroma")
+chroma_path = os.path.join(os.path.dirname(__file__), CHROMA_DIR)
 if not os.path.exists(chroma_path) or not os.listdir(chroma_path):
     logger.info("Chroma DB is empty, inserting documents...")
     insert_documents()
