+++ src/backend/question_generator_base.py
@@ -1,7 +1,7 @@
 from langchain_ollama import OllamaLLM
 from langchain_core.prompts import ChatPromptTemplate
-from .template import create_question_template
-from .vectorstore import load_vectorstore
+from template import create_question_template
+from vectorstore import load_vectorstore
 
 
 class QuestionGeneratorBase:
