+++ src/backend/rag_pipeline.py
@@ -1,7 +1,7 @@
-from .local_loader import load_documents
-from .splitter import smart_split_documents
-from .vectorstore import store_documents
-from .Flashcards_llm_ollama import FlashCards
+from local_loader import load_documents
+from splitter import smart_split_documents
+from vectorstore import store_documents
+from Flashcards_llm_ollama import FlashCards
 import os
 
 # --- CONFIG ---
